Added Custom names For Url and made ui changes
Added B4

+++ shortner/views.py
@@ -17,35 +17,68 @@ def index2(request):
     return redirect('/create')
 
 def create(request):
+    context= {}
     base_url = os.environ['SERVER']
-    base_scheme = request.scheme
     if request.method == 'POST':
         _url = request.POST.get('url')
         if Urls.objects.filter(url=_url):
             url = Urls.objects.filter(url=_url)[0]
             complete_url = f'{base_url}{url.slug}' 
             print(complete_url,'.........................')
-            return HttpResponse(f'Your Url is : <a href={complete_url} target=_blank>{complete_url}</a>' )
+            # return HttpResponse(f'Your Url is : <a href={complete_url} target=_blank>{complete_url}</a>' )
+            context['url'] = complete_url
+            context['status'] = '1'
+            return render(request,'created.html', context)
         else:
-            db_url = Urls()
-            db_url.slug = generate('abcdefghijklmnopqrstuvwxyz1234567890',5)
-            db_url.url = request.POST.get('url')
-            db_url.save()
-            complete_url = f'{base_url}{db_url.slug}'
-            return HttpResponse(f'Your Url is : <a href={complete_url} target=_blank>{complete_url}</a>' )
+            if request.POST.get('myslug'):
+                myslug = request.POST.get('myslug')
+                if Urls.objects.filter(slug=myslug):
+                    url = Urls.objects.filter(slug=myslug)[0]
+                    complete_url = url.url 
+                    print(complete_url,'.........................')
+                    # return HttpResponse(f'This URL <a href={complete_url} target=_blank>{complete_url}</a> already exists with name: {myslug}' )
+                    context['url'] = complete_url
+                    context['myslug'] = url.slug
+                    context['status'] = '2'
+                    return render(request,'created.html', context)
+                else:
+                    db_url = Urls()
+                    db_url.slug = request.POST.get('myslug')
+                    db_url.url = request.POST.get('url')
+                    db_url.save()
+                    complete_url = f'{base_url}{db_url.slug}'
+                    # return HttpResponse(f'Your Url is : <a href={complete_url} target=_blank>{complete_url}</a>' )    
+                    context['url'] = complete_url
+                    context['myslug'] = db_url.slug
+                    context['status'] = '3'
+                    return render(request,'created.html', context)
+            else:
+                db_url = Urls()
+                db_url.slug = generate('abcdefghijklmnopqrstuvwxyz1234567890',6)
+                db_url.url = request.POST.get('url')
+                db_url.save()
+                complete_url = f'{base_url}{db_url.slug}'
+                context['url'] = complete_url
+                ontext['status'] = '4'
+                return render(request,'created.html', context)
+                # return HttpResponse(f'Your Url is : <a href={complete_url} target=_blank>{complete_url}</a>' )
     else:
         return render(request,'index.html')
 
 def list_short(request):
+    base_url = os.environ['SERVER']
     context = {}
-    if request.method == 'POST':
-        url = Urls.objects.get(slug=request.POST.get('slug'))
-        url.delete()    
-        return redirect('/list')
+    myslug = request.POST.get('slug')
+    if myslug:
+        if request.method == 'POST':
+            url = Urls.objects.filter(slug=myslug)[0]
+            url.delete()    
+            return redirect('/list/')
     else:
         url_list = Urls.objects.all()
         context['list'] = url_list
+        context['server'] = base_url
         
-        return render(request,'list.html',context)
+    return render(request,'list.html',context)
